update: ARM_WEAP.DAT, HERC_INF.DAT, INIT_[herc].DAT mapped and mostly understood.
+ TODO - transformers for these.

+++ src/main/java/com/mech/works/data/file/dyn/DynamixBitmapArray.java
@@ -20,8 +20,8 @@ public class DynamixBitmapArray extends DataFile{
 	
 	private LinkedHashSet<DynamixBitmap> images;
 	
-	private int arrayRow;
-	private int arrayCols;
+	private short arrayRow;
+	private short arrayCols;
 
 	private DynamixPalette palette;
 	
@@ -36,19 +36,19 @@ public class DynamixBitmapArray extends DataFile{
 		this.images = images;
 	}
 
-	public int getArrayRow() {
+	public short getArrayRow() {
 		return arrayRow;
 	}
 
-	public void setArrayRow(int arrayRow) {
+	public void setArrayRow(short arrayRow) {
 		this.arrayRow = arrayRow;
 	}
 
-	public int getArrayCols() {
+	public short getArrayCols() {
 		return arrayCols;
 	}
 
-	public void setArrayCols(int arrayCols) {
+	public void setArrayCols(short arrayCols) {
 		this.arrayCols = arrayCols;
 	}
